requirements.txt
websockets~=14.2
pvorca~=1.1.0
pvspeaker~=1.0.4
picollm ~=1.2.4
ollama~=0.4.7
pvcheetah~=2.1.2
pvrecorder~=1.2.4
python-dotenv~=1.0.1